REFACTOR: 프로젝트 구조 수정 * 디렉토리 수정
