+++ src/main/java/f_thon/godlifebingo/common/security/dto/GoogleOAuth2Response.java
@@ -1,4 +1,4 @@
-package f_thon.godlifebingo.security.dto;
+package f_thon.godlifebingo.common.security.dto;
 
 import java.util.Map;
