+++ src/main/java/f_thon/godlifebingo/core/bingo/BingoService.java
@@ -2,8 +2,8 @@ package f_thon.godlifebingo.core.bingo;
 
 import f_thon.godlifebingo.core.bingo.dto.BingoCreateRequest;
 import f_thon.godlifebingo.core.bingo.dto.BingoGetResponse;
-import f_thon.godlifebingo.core.cell.BingoCellInfo;
-import f_thon.godlifebingo.core.cell.BingoCellResponse;
+import f_thon.godlifebingo.core.cell.dto.BingoCellInfo;
+import f_thon.godlifebingo.core.cell.dto.BingoCellResponse;
 import f_thon.godlifebingo.core.bingo.dto.BingoListResponse;
 import f_thon.godlifebingo.core.bingo.dto.BingoResponse;
 import f_thon.godlifebingo.core.bingo.dto.BingoRow;
@@ -13,9 +13,7 @@ import f_thon.godlifebingo.core.cell.Cell;
 import f_thon.godlifebingo.core.cell.CellRepository;
 import f_thon.godlifebingo.core.godlife.GodLifeRepository;
 import f_thon.godlifebingo.core.users.Users;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
