+++ src/main/java/f_thon/godlifebingo/core/cell/dto/BingoCellResponse.java
@@ -1,7 +1,8 @@
-package f_thon.godlifebingo.core.cell;
+package f_thon.godlifebingo.core.cell.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
+import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -12,11 +13,6 @@ import lombok.NoArgsConstructor;
 @JsonInclude(Include.NON_NULL)
 @AllArgsConstructor
 @NoArgsConstructor
-public class BingoCellInfo {
-
-    private String title;
-    private Long id;
-    private Double achievementRate;
-    private int row;
-    private int col;
+public class BingoCellResponse {
+    private List<BingoCellInfo> bingoCells;
 }
