+++ src/main/java/f_thon/godlifebingo/core/godlife/dto/GodLifeStatistic.java
@@ -1,4 +1,4 @@
-package f_thon.godlifebingo.core.godlife;
+package f_thon.godlifebingo.core.godlife.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
